test fix: only delete files that exist

+++ app/tests/__init__.py
@@ -2,7 +2,8 @@ import os
 import logging
 
 TEST_LOG_PATH = 'logs/chesspy.test.log'
-os.unlink(TEST_LOG_PATH)
+if os.path.exists(TEST_LOG_PATH):
+    os.unlink(TEST_LOG_PATH)
 logging.basicConfig(filename=TEST_LOG_PATH, encoding='utf-8', level=logging.DEBUG)
 
 from .test_san import *
